server/requirements.txt
bilibili-api-python
fastapi
uvicorn
httpx
aiowebsocket
apscheduler
pydantic
pyinstaller